Fix some trivial scaling problems
I used to scale everything to be square.  That was fine for testing, but
not so great for actual use.  This makes it all look better when the
input isn't square.  It also scales the width of the line properly.

And in honor of this actually being useful (it's installed on Tove's
machine now), let's sign off on it all.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>