Add a 'density' argument for thread thickness
I like the default thin thread effect, but using a density of 2.5 or so
seems to be a bit more realistic on my test-case.  Which may be more
about the test-case than anything else, but whatever.

So with this, you can do something like

	pesconvert -d 2.5 -s 512 Jan_heartsdelight.pes s512.png

to generate a 512-pixel sized version of the Brother example file.

Making the "thread" thicker also meant that the line join matters a lot
more.  A miter join (the cairo default) is not how thread works and
makes it all look very jagged.  So make it use a round join instead.

The line cap could probably be a butt cap, but I set it to round too.  I
doubt anybody will see any difference, since thread ends aren't common.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>